# swagger-check.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Volvid API",
    "version": "1.0.0",
    "description": "API for Volvid pets platform"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Crear cuenta de usuario",
        "description": "Registra un nuevo usuario en Volvid y devuelve un Token JWT para la app.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cuenta creada exitosamente. Token generado."
          },
          "400": {
            "description": "Faltan datos requeridos o las contraseñas no coinciden."
          },
          "409": {
            "description": "El correo ya está registrado en el sistema."
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Iniciar sesión",
        "description": "Valida el correo y contraseña ingresados por el usuario y genera un Token JWT.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso, se retorna la información del usuario y token."
          },
          "400": {
            "description": "Datos faltantes en el cuerpo de la petición."
          },
          "401": {
            "description": "Credenciales inválidas (contraseña incorrecta o correo no existe)."
          }
        }
      }
    },
    "/api/pets": {
      "get": {
        "summary": "Obtener todas las mascotas",
        "tags": [
          "Pets"
        ],
        "responses": {
          "200": {
            "description": "Lista de mascotas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear una nueva mascota",
        "tags": [
          "Pets"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pet"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mascota creada"
          }
        }
      }
    },
    "/api/pets/user/{owner_name}": {
      "get": {
        "summary": "Obtener mascotas filtradas por usuario (dueño)",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "owner_name",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Nombre del dueño de la mascota"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de mascotas que pertenecen a este usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Pet"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/pets/{id}": {
      "get": {
        "summary": "Obtener mascota por ID",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID de la mascota"
          }
        ],
        "responses": {
          "200": {
            "description": "Datos de la mascota"
          },
          "404": {
            "description": "Mascota no encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar una mascota",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mascota actualizada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar una mascota",
        "tags": [
          "Pets"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Mascota eliminada"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegistrationRequest": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre completo Ej. Juan Pérez"
          },
          "email": {
            "type": "string",
            "description": "Correo electrónico tu@ejemplo.com"
          },
          "phone": {
            "type": "string",
            "description": "Número de teléfono +34 000..."
          },
          "role": {
            "type": "string",
            "description": "Rol del usuario (ej. owner, provider)",
            "default": "owner"
          },
          "accepted_terms": {
            "type": "boolean",
            "description": "Aceptación de términos y condiciones y políticas de privacidad"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Contraseña segura"
          },
          "confirmPassword": {
            "type": "string",
            "format": "password",
            "description": "Confirmación opcional de contraseña"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "Correo electrónico registrado"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Contraseña"
          }
        }
      },
      "Pet": {
        "type": "object",
        "required": [
          "name",
          "species"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID autogenerado"
          },
          "name": {
            "type": "string",
            "description": "Nombre de la mascota"
          },
          "species": {
            "type": "string",
            "description": "Especie (Dog, Cat, etc.)"
          },
          "breed": {
            "type": "string",
            "description": "Raza"
          },
          "age": {
            "type": "integer",
            "description": "Edad en años"
          },
          "owner_name": {
            "type": "string",
            "description": "Nombre del dueño"
          }
        }
      }
    }
  },
  "tags": []
}